% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occMod.R
\name{occMod}
\alias{occMod}
\title{Run occupancy models in \code{rapidPop}.}
\usage{
occMod(
  input_file = NULL,
  input_file_type = ".txt",
  nm1 = "day01",
  nmF = "day12",
  parameter = FALSE,
  nm_parameter,
  digits = 3,
  shiny = FALSE
)
}
\arguments{
\item{input_file}{The path to and name of your input file. This package includes an
\code{example_input_file.txt} that might be helpful.}

\item{input_file_type}{The file type (or extenstion). Options are c(".txt", ".csv")}

\item{nm1}{The name of the first column containing occupancy observation data. 
Your \code{input_file} must have some columns contatining information about 
whether there were animals in your observations. These columns can be either 
0s and 1s or the number of animals in the image. \code{nm1} is the name (or header)
of the first column containing these data.}

\item{nmF}{The name of the final column containing occupancy data.}

\item{parameter}{logical. Do you want to model the effect of a parameter on occupancy?
For example, do you have a column indicating if these observations came from
before or after control operations? If so, you could model the effect of this
parameter on occupancy.}

\item{nm_parameter}{The name of the column for your \code{parameter}. If you
are modeling a paramter (\code{parameter = TRUE}), this is the column name
for the paramter you want to use.}

\item{digits}{The number of digits to round to. This number is passed to the
function \code{round}}
}
\description{
Run occupancy models in \code{rapidPop}.
}
