% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{A wrapper function to run Shiny Apps from \code{rapidPop}.}
\usage{
runShiny(app = "occMod")
}
\arguments{
\item{app}{The name of the app you want to run. The options is `occMod`}
}
\description{
A wrapper function to run Shiny Apps from \code{rapidPop}.
}
